# cmd/api/api.go
package main

import (
	er "golang-http/internal/errors"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	httpSwagger "github.com/swaggo/http-swagger/v2"

	_ "golang-http/docs" // docs is generated by Swag CLI, you have to import it.
)

var (
	errorSpecifyAPIVer = "Please specify the version of the api."
	incorrectPath      = "The path provided is not valid. Maybe it could be a wrong HTTP Method?"
)

func (s *Server) RegisterRoutes() http.Handler {
	r := chi.NewRouter()

	r.Use(middleware.RequestID) //adds a unique counter for each request
	r.Use(middleware.RealIP)
	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer) // recovers from panics, logs the panic, and returns a 500 status if possible

	r.Use(cors.Handler(cors.Options{
		AllowedOrigins:   []string{"https://*", "http://*"},
		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type"},
		AllowCredentials: true,
		MaxAge:           300,
	}))

	r.Route("/v1", func(r chi.Router) {
		r.Route("/auth", func(r chi.Router) {
			r.Post("/user", s.registerUser)
			r.Post("/token", s.loginUser)
		})

		//! from this point on, every route is protected
		r.Group(func(r chi.Router) {
			r.Use(s.AuthMiddleware)

			r.Get("/health", s.HealthHandler)

			//  routes for /v1/user
			r.Route("/user", func(r chi.Router) {
				r.Get("/", s.getOwnUser)
				r.Get("/{id}", s.getUser)
			})
		})

		r.NotFound(func(w http.ResponseWriter, r *http.Request) {
			s.WriteJSONError(w, http.StatusNotFound, &er.ErrorsStruct{Message: incorrectPath})
		})
	})

	r.Get("/swagger/*", httpSwagger.Handler(
		httpSwagger.URL("/swagger/doc.json")),
	)

	r.NotFound(func(w http.ResponseWriter, r *http.Request) {
		s.WriteJSONError(w, http.StatusNotFound, &er.ErrorsStruct{Message: errorSpecifyAPIVer})
	})

	return r
}